requirements.txt
streamlit==1.25.0
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
weave
boto3==1.28.38
litellm
tiktoken==0.4.0
aiohttp==3.8.5
pydantic==2.3.0